# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FlowForge API",
    "description": "From intent to execution. The LLM plans, the state machine executes, the human stays in control.",
    "version": "0.1.0"
  },
  "paths": {
    "/workflows": {
      "post": {
        "summary": "Create Workflow",
        "operationId": "create_workflow_workflows_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkflowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateWorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workflows/{workflow_id}": {
      "get": {
        "summary": "Get Workflow",
        "operationId": "get_workflow_workflows__workflow_id__get",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workflows/{workflow_id}/advance": {
      "post": {
        "summary": "Advance",
        "operationId": "advance_workflows__workflow_id__advance_post",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceWorkflowRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvanceWorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workflows/{workflow_id}/documents": {
      "post": {
        "summary": "Upload Document",
        "operationId": "upload_document_workflows__workflow_id__documents_post",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_workflows__workflow_id__documents_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workflows/{workflow_id}/audit": {
      "get": {
        "summary": "List Audit",
        "operationId": "list_audit_workflows__workflow_id__audit_get",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "summary": "Root",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Root  Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AdvanceWorkflowRequest": {
        "properties": {
          "user_input": {
            "additionalProperties": true,
            "type": "object",
            "title": "User Input",
            "description": "Values for the current user_input state, keyed by required_data field."
          },
          "approval": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval",
            "description": "Explicit approve/deny decision for the current human_approval state."
          },
          "acknowledge": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledge",
            "description": "Acknowledge a warning gate (e.g. a low-confidence validation issue)."
          },
          "confirm": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirm",
            "description": "Generic confirmation for user_confirmed transitions."
          },
          "document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id",
            "description": "Optional reference to a specific uploaded document."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AdvanceWorkflowRequest",
        "description": "Payload accepted by POST /workflows/{id}/advance.\n\nOnly the field relevant to the current gated state is consumed; other fields are\nignored by the deterministic executor."
      },
      "AdvanceWorkflowResponse": {
        "properties": {
          "workflow_id": {
            "type": "string",
            "title": "Workflow Id"
          },
          "status": {
            "$ref": "#/components/schemas/WorkflowStatus"
          },
          "goal": {
            "type": "string",
            "title": "Goal"
          },
          "current_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current State",
            "description": "id of the active state."
          },
          "last_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Message",
            "description": "Human-readable assistant message for the current state."
          },
          "needs": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Needs",
            "description": "Normalized gate label: user_input | document_upload | approval | action."
          },
          "progress": {
            "$ref": "#/components/schemas/WorkflowProgress"
          },
          "states": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "States",
            "description": "Serialized state nodes for the graph view."
          },
          "collected_documents": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Collected Documents",
            "description": "Classifications of documents received so far."
          },
          "validation": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation",
            "description": "Latest cross-document validation result, if any."
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Status message for this step."
          },
          "completed": {
            "type": "boolean",
            "title": "Completed",
            "description": "True once a terminal state is reached.",
            "default": false
          },
          "events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Events",
            "description": "Audit events emitted by this step."
          }
        },
        "type": "object",
        "required": [
          "workflow_id",
          "status",
          "goal",
          "progress",
          "states"
        ],
        "title": "AdvanceWorkflowResponse",
        "description": "Result of one advance step: the updated detail plus what the executor did."
      },
      "AuditListResponse": {
        "properties": {
          "workflow_id": {
            "type": "string",
            "title": "Workflow Id"
          },
          "events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "workflow_id"
        ],
        "title": "AuditListResponse"
      },
      "Body_upload_document_workflows__workflow_id__documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_workflows__workflow_id__documents_post"
      },
      "CreateWorkflowRequest": {
        "properties": {
          "goal": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Goal",
            "description": "Natural-language goal the workflow should accomplish."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "goal"
        ],
        "title": "CreateWorkflowRequest"
      },
      "CreateWorkflowResponse": {
        "properties": {
          "workflow_id": {
            "type": "string",
            "title": "Workflow Id",
            "description": "Server-assigned workflow identifier."
          },
          "status": {
            "$ref": "#/components/schemas/WorkflowStatus"
          },
          "workflow": {
            "additionalProperties": true,
            "type": "object",
            "title": "Workflow",
            "description": "The validated workflow definition (see docs/workflow-engine.md)."
          }
        },
        "type": "object",
        "required": [
          "workflow_id",
          "status",
          "workflow"
        ],
        "title": "CreateWorkflowResponse"
      },
      "DocumentUploadResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "classification": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classification",
            "description": "Detected document category."
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence",
            "description": "Classification confidence in [0, 1]."
          },
          "extracted_fields": {
            "additionalProperties": true,
            "type": "object",
            "title": "Extracted Fields",
            "description": "Extracted fields, each with value/confidence/source_text."
          },
          "validation_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Status",
            "description": "Per-document status: pass | needs_review | block."
          },
          "issues": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Issues"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "filename"
        ],
        "title": "DocumentUploadResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkflowDetailResponse": {
        "properties": {
          "workflow_id": {
            "type": "string",
            "title": "Workflow Id"
          },
          "status": {
            "$ref": "#/components/schemas/WorkflowStatus"
          },
          "goal": {
            "type": "string",
            "title": "Goal"
          },
          "current_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current State",
            "description": "id of the active state."
          },
          "last_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Message",
            "description": "Human-readable assistant message for the current state."
          },
          "needs": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Needs",
            "description": "Normalized gate label: user_input | document_upload | approval | action."
          },
          "progress": {
            "$ref": "#/components/schemas/WorkflowProgress"
          },
          "states": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "States",
            "description": "Serialized state nodes for the graph view."
          },
          "collected_documents": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Collected Documents",
            "description": "Classifications of documents received so far."
          },
          "validation": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation",
            "description": "Latest cross-document validation result, if any."
          }
        },
        "type": "object",
        "required": [
          "workflow_id",
          "status",
          "goal",
          "progress",
          "states"
        ],
        "title": "WorkflowDetailResponse"
      },
      "WorkflowProgress": {
        "properties": {
          "completed": {
            "type": "integer",
            "title": "Completed",
            "description": "Number of states completed."
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of states in the workflow."
          },
          "ratio": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Ratio",
            "description": "completed / total."
          }
        },
        "type": "object",
        "required": [
          "completed",
          "total",
          "ratio"
        ],
        "title": "WorkflowProgress"
      },
      "WorkflowStatus": {
        "type": "string",
        "enum": [
          "created",
          "in_progress",
          "blocked",
          "generation_failed",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "WorkflowStatus"
      }
    }
  }
}